pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-ooxml-engine"
version = "0.1.0"
description = "Model-agnostic native OOXML execution engine for PPTX operations."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Xi Zhao" }
]
dependencies = [
  "python-pptx>=1.0.2",
  "pydantic>=2.8.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "build>=1.2.0"
]

[project.scripts]
pptx-ooxml-engine = "pptx_ooxml_engine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pptx_ooxml_engine = ["schemas/*.json"]

[tool.pytest.ini_options]
pythonpath = ["src"]
